Split audiomgr and audiosfx files
