+++ src/audiomgr.h
@@ -1,5 +1,5 @@
-#ifndef _UNKNOWN_003260_H_
-#define _UNKNOWN_003260_H_
+#ifndef _AUDIOMGR_H_
+#define _AUDIOMGR_H_
 
 #include "types.h"
 #include "macros.h"
@@ -156,13 +156,6 @@ typedef struct
     char        *ptr;
 } AMDMABuffer;
 
-/* Unknown Size */
-/*typedef struct unk800DC6B0 {
-    ALLink node;
-    ALLink_s * unk08;
-} unk800DC6B0;
-*/
-
 typedef struct audioMgrConfig_s{
     u32 unk00;
     u32 unk04;
@@ -177,59 +170,20 @@ typedef struct {
     AMDMABuffer *firstFree;
 } AMDMAState;
 
-extern u32 audFrameCt;
-extern u32 nextDMA;
-extern u32 curAcmdList;
-extern s8 gAntiPiracyAudioFreq;
-extern s32 gFunc80019808Checksum;
-extern s32 gFunc80019808Length;
-extern s32 D_800DC6A0;
-extern s32 D_800DC6A4;
-extern s32 D_800DC6A8;
-extern s32 D_800DC6AC;
-
-extern ALEventQueue *D_800DC6B0;
-extern s32 D_800DC6B4;
-extern unk800DC6BC_40* D_800DC6B8;
-extern unk800DC6BC* gAlSndPlayer;
-
-extern s32 sfxVolumeSlider;
-extern s32 D_800DC6C4;
 extern s8 rspF3DDKRBootStart[];
 extern s8 rspF3DDKRDramStart[];
 extern s8 aspMainDataStart[];
 extern s8 aspMainTextStart[];
-extern s32 gAudioCmdLen;
 
-Acmd *alAudioFrame(Acmd *cmdList, s32 *cmdLen, s16 *outBuf, s32 outLen);
-void *alHeapDBAlloc(u8 *file, s32 line, ALHeap *hp, s32 num, s32 size); //lib/src/al
-void alEvtqNew(ALEventQueue *evtq, ALEventListItem *items, s32 itemCount); //lib/src/unknown_0C9C90.c
-ALMicroTime alEvtqNextEvent(ALEventQueue *evtq, ALEvent *evt); //lib/src/unknown_0C9C90.c
-void alSynAddPlayer(ALSynth *drvr, ALPlayer *client); //lib/src/unknown_0C9C90.c
-void alEvtqPostEvent(ALEventQueue *evtq, ALEvent *evt, ALMicroTime delta); //lib/src/unknown_0C9C90.c Non Matching
+extern Acmd *alAudioFrame(Acmd *cmdList, s32 *cmdLen, s16 *outBuf, s32 outLen);
+extern void *alHeapDBAlloc(u8 *file, s32 line, ALHeap *hp, s32 num, s32 size); //lib/src/al
+extern void alEvtqNew(ALEventQueue *evtq, ALEventListItem *items, s32 itemCount); //lib/src/unknown_0C9C90.c
+extern ALMicroTime alEvtqNextEvent(ALEventQueue *evtq, ALEvent *evt); //lib/src/unknown_0C9C90.c
+extern void alSynAddPlayer(ALSynth *drvr, ALPlayer *client); //lib/src/unknown_0C9C90.c
+extern void alEvtqPostEvent(ALEventQueue *evtq, ALEvent *evt, ALMicroTime delta); //lib/src/unknown_0C9C90.c Non Matching
 
 void audioStartThread(void);
 void audioStopThread(void);
-void set_sfx_volume_slider(u32 arg0);
-s32 sfxGetVolumeSlider(void);
-void set_sound_channel_count(s32 numChannels);
-void alSndPNew(audioMgrConfig *c);
-void func_80004604(u8 *arg0, u8 arg1);
-u8 func_8000461C(u8 *arg0);
-void func_80004638(ALBank *bnk, s16 sndIndx, s32 arg2);
-INCONSISTENT void func_8000488C();
-void func_800049D8(void);
-void func_800049F8(s32 soundMask, s16 type, u32 volume);
-u16 get_sound_channel_volume(u8 arg0);
-void set_sound_channel_volume(u8 channel, u16 volume);
-u16 func_800042CC(u16 *lastAllocListIndex, u16 *lastFreeListIndex);
-void func_8000410C(ALSoundState *state);
 void amCreateAudioMgr(ALSynConfig *c, OSPri pri, OSSched *audSched);
-
-// Non Matching
-ALMicroTime  _sndpVoiceHandler(void *node);
-void func_80004668(ALBank *bnk, s16 sndIndx, u8, s32);
-void func_800048D8(s32);
-void func_80004520(ALSoundState *);
 
 #endif
